Ma'lumot o'chirishda agar boshqa tablitsalarga bog'langan columnlar bo'lsa o'chirish mumkin emasligi haqidagi toastr chiqadi udalit qilishda

+++ app/Http/Controllers/SizesController.php
@@ -86,6 +86,9 @@ class SizesController extends Controller
     public function destroy(string $id)
     {
         $model = Sizes::find($id);
+        if($model->warehouse){
+            return redirect()->back()->with('error', translate('You cannot delete this size because here is product associated with this size.'));
+        }
         $model->delete();
         return redirect()->route('size.index')->with('status', translate('Successfully deleted'));
     }
