+++ app/Http/Controllers/SubCategoryController.php
@@ -110,6 +110,9 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $model = Category::where('step', 1)->find($id);
+        if($model->product) {
+            return redirect()->back()->with('error', translate('You cannot delete this category because here is product associated with this size.'));
+        }
         $model->delete();
         return redirect()->route('subcategory.index', $model->category->id)->with('status', translate('Successfully deleted'));
     }
